+++ src/EventLogExpert/Components/DetailsPane.razor
@@ -1,4 +1,5 @@
-﻿@inherits FluxorComponent
+﻿@using EventLogExpert.Services;
+@inherits FluxorComponent
 
 @inject IState<EventLogState> EventLogState
 @inject IState<SettingsState> SettingsState
