+++ src/EventLogExpert/Components/DetailsPane.razor.cs
@@ -2,6 +2,7 @@
 // // Licensed under the MIT License.
 
 using EventLogExpert.Library.Models;
+using EventLogExpert.Services;
 using EventLogExpert.Store.EventLog;
 using Fluxor;
 using Microsoft.AspNetCore.Components;
