+++ src/EventLogExpert/Components/EventTable.razor
@@ -1,4 +1,5 @@
-﻿@inherits FluxorComponent
+﻿@using EventLogExpert.Services;
+@inherits FluxorComponent
 
 @inject IState<EventLogState> EventLogState
 @inject IState<FilterPaneState> FilterPaneState
